+++ src/app/chat/ChatBot/style.ts
@@ -1,15 +1,11 @@
 import { createStyles } from 'antd-style';
 
 const useStyles = createStyles(({ css }) => ({
-  chat: css`
+  chatBot: css`
     display: flex;
-    flex-direction: row;
+    flex-direction: column;
     flex-grow: 1;
 
-    width: 100%;
-    height: 100%;
-  `,
-  content: css`
     width: 100%;
     height: 100%;
   `,
