+++ src/app/chat/VirtualIdol/Dialog/index.tsx
