+++ src/app/chat/VirtualIdol/Dialog/style.ts
