+++ src/app/chat/VirtualIdol/Docker/ToolBar/index.tsx
