+++ src/app/chat/VirtualIdol/Docker/ToolBar/style.ts
