+++ src/app/chat/VirtualIdol/index.tsx
@@ -4,6 +4,8 @@ import AgentViewer from '@/features/AgentViewer';
 import ImageViewer from '@/features/ImageViewer';
 import { useSessionStore } from '@/store/session';
 
+import Dialog from './Dialog';
+import Docker from './Docker';
 import { useStyles } from './style';
 
 const VirtualIdol = () => {
@@ -11,7 +13,13 @@ const VirtualIdol = () => {
   const [viewerMode] = useSessionStore((s) => [s.viewerMode]);
 
   return (
-    <div className={styles.content}>{viewerMode === true ? <AgentViewer /> : <ImageViewer />}</div>
+    <div className={styles.container}>
+      <div style={{ flex: 1, overflow: 'hidden', position: 'relative' }}>
+        {viewerMode === true ? <AgentViewer /> : <ImageViewer />}
+      </div>
+      <Dialog />
+      <Docker />
+    </div>
   );
 };
