+++ src/app/chat/VirtualIdol/style.ts
@@ -1,7 +1,11 @@
 import { createStyles } from 'antd-style';
 
 export const useStyles = createStyles(({ css }) => ({
-  content: css`
+  container: css`
+    display: flex;
+    flex-direction: column;
+    flex-grow: 1;
+
     width: 100%;
     height: 100%;
   `,
