+++ src/features/ChatHeader/index.tsx
@@ -1,6 +1,7 @@
 import AgentMeta from '@/components/AgentMeta';
 import { sessionSelectors, useSessionStore } from '@/store/session';
 
+import ViewerMode from './ViewerMode';
 import { useStyles } from './style';
 
 const Header = () => {
@@ -10,6 +11,7 @@ const Header = () => {
   return (
     <div className={styles.header}>
       <AgentMeta meta={currentAgent?.meta} />
+      <ViewerMode key={'viewer'} />
     </div>
   );
 };
