+++ src/features/ChatInfo/index.tsx
@@ -11,8 +11,6 @@ import { agentListSelectors, useAgentStore } from '@/store/agent';
 import { useConfigStore } from '@/store/config';
 import { sessionSelectors, useSessionStore } from '@/store/session';
 
-import ViewerMode from '../ViewerMode';
-
 const useStyles = createStyles(({ css, token }) => ({
   content: css`
     display: flex;
@@ -64,7 +62,6 @@ const Header = () => {
             编辑
           </Button>,
           <Voice key={'voice'} />,
-          <ViewerMode key={'viewer'} />,
         ]}
         agent={currentAgent}
       />
