+++ src/layout/Background/style.ts
