+++ src/layout/Header/index.tsx
@@ -35,10 +35,6 @@ const Header = (props: Props) => {
               label: '聊天',
             },
             {
-              key: HeaderNavKey.Home,
-              label: '互动',
-            },
-            {
               key: HeaderNavKey.Market,
               label: '发现',
             },
