+++ src/layout/type.ts
@@ -1,6 +1,5 @@
 export enum HeaderNavKey {
   Chat = 'chat',
-  Home = 'home',
   Market = 'market',
   My = 'my',
 }
